model_api/requirements.txt.txt
uvicorn
fastapi
torch
numpy
pillow
python-multipart